refactor(agents,cli): extract plan-converter.ts and help-text.ts
Round 5 architecture deepening — Candidates #3 + #4 in parallel.

1. plan-converter.ts (new ~120 lines) — extracted planToBuildSteps,
   buildFlatFormSteps, parsePlanToSteps + BuildStep/BuildAction types
   from build-agent.ts into a focused plan-parsing module.

2. help-text.ts (new ~140 lines) — extracted showHelp() function
   from main.tsx (405 lines → ~260 lines) into its own content module.

3. build-agent.ts — now imports parsePlanToSteps as parseSteps from
   plan-converter.ts. Re-exports BuildStep, BuildAction, and
   parsePlanToSteps for backward compat. Removed 78 lines of dead
   code.

4. step-executor.ts — imports BuildAction/BuildStep from
   plan-converter.ts instead of build-agent.ts (no circular deps).

5. main.tsx — imports showHelp from help-text.ts.

All 600+ tests pass. Lint clean (241 files).
